refactor: DecoderとDecodedFrameを別モジュールに分離
